refactor(external_service): gather http client
