+++ src/application/core/external_service/auth_client.py
@@ -5,7 +5,7 @@ from typing import Literal
 from aiohttp import ClientSession
 
 from application.core.enums import ResponseCode
-from application.core.utils.http_client import BaseHttpClient
+from application.core.external_service.http_client import BaseHttpClient
 from application.core.exceptions.token import (
     TokenDecodeException,
     TokenExpireException,
