+++ src/application/core/external_service/http_client.py
